Extract DockerFileParser to its own file
Changes are needed in this file for 212+ so go ahead and isolate it to make source set overrides easier
