+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/ecs/execution/docker/DockerfileParserTest.kt
@@ -1,7 +1,7 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.ecs.execution
+package software.aws.toolkits.jetbrains.services.ecs.execution.docker
 
 import com.intellij.docker.dockerFile.DockerFileType
 import com.intellij.docker.dockerFile.DockerLanguage
@@ -158,7 +158,7 @@ class DockerfileParserTest {
                 COPY . .
             """.trimIndent()
         )
-        val directory = file.parent.path.normalizeDirectory(matchPlatform = true)
+        val directory = file.parent.path + "/"
         runInEdtAndWait {
             assertThat(sut.parse(file)).isEqualTo(
                 DockerfileDetails(
